refactor move limiting_factor to sqllab package
